# backend/internal/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: RallyUp API
  description: RallyUp Badminton Club management API — member registration, session scheduling, RSVP tracking, and notifications.
  version: 1.0.0
servers:
  - url: /api
    description: API root
...
security:
  - bearerAuth: []
paths:
...
  /accounts:
    get:
      tags: [Ledger]
      summary: List every member's balance
      description: >
        Visible to all approved members. The club already worked this way in
        Splitwise and is comfortable with it. Amounts are integer cents.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PlayerBalance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /accounts/me:
    get:
      tags: [Ledger]
      summary: The caller's own balance
      description: >
        Deliberately minimal — this backs the balance chip in the header and is
        requested on nearly every screen. The state is derived server-side so
        there is one definition of "running low" rather than two that drift.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Balance and state
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance_cents:
                    type: integer
                    format: int64
                  state:
                    type: string
                    enum: [ok, low, negative]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/me/entries:
    get:
      tags: [Ledger]
      summary: The caller's own itemised history
      description: >
        Newest first. Each row carries the balance it produced, computed over the
        member's whole history and only then paged, so page two does not restart
        the arithmetic from zero.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/LedgerEntryView"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/entries:
    get:
      tags: [Ledger]
      summary: Filtered member ledger history
      description: >
        Newest first. Each row carries the balance it produced, computed over the
        member's whole history before filtering or paging. Defaults to the caller.
        The all scope includes inactive player accounts, without bank/surplus legs.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all], default: mine } }
        - { name: type, in: query, schema: { type: string, enum: [all, topup], default: all } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/LedgerEntryView"
                  total:
                    type: integer
        "400":
          description: Invalid scope or type
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/activity:
    get:
      tags: [Ledger]
      summary: Ledger activities with complete game splits
      description: >
        One expandable activity per game. Imported regular and extra-hour rows on
        the same play date form one game. Native settlement transactions stay
        distinct. Mine selects games the caller participated in while retaining
        every account share. Grouping happens before paging. Non-game account
        entries keep their full running balances. The total counts activities.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all], default: mine } }
        - { name: type, in: query, schema: { type: string, enum: [all, topup], default: all } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/LedgerActivityView"
                  total:
                    type: integer
        "400":
          description: Invalid scope or type
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/transactions/topup:
    post:
      tags: [Ledger]
      summary: Record money a member paid into the club
      description: Bank up, that member up. The club now holds their money and owes them its value.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashRequest"
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/transactions/withdrawal:
    post:
      tags: [Ledger]
      summary: Record the club paying a member back
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashRequest"
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/transactions/court-credit:
    post:
      tags: [Ledger]
      summary: Record topping up the venue account
      description: >
        Bank down, court credit up. No player balance moves — this is the club
        swapping one asset for another, which is exactly why a single club
        balance cannot answer whether the club is square.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetPurchaseRequest"
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/transactions/shuttle-purchase:
    post:
      tags: [Ledger]
      summary: Record shuttles bought
      description: >
        Both the value and the count go into stock. Buying at a new price blends
        automatically and never revalues what is already in the bag.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetPurchaseRequest"
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/transactions/opening-balances:
    post:
      tags: [Ledger]
      summary: Seed the ledger at go-live
      description: >
        Accepted once. Surplus is not supplied by the caller but computed as
        whatever makes the books balance.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpeningBalancesRequest"
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "409":
          $ref: "#/components/responses/LedgerConflict"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/transactions/{id}/reverse:
    post:
      tags: [Ledger]
      summary: Reverse a transaction
      description: >
        Posts the exact negation of the original. This is the only way to undo
        anything — there is deliberately no edit or delete endpoint, and both the
        original and its reversal stay visible.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/TransactionCreated"
        "409":
          $ref: "#/components/responses/LedgerConflict"
        "403":
          $ref: "#/components/responses/Forbidden"
  /position:
    get:
      tags: [Ledger]
      summary: The club's true position
      description: >
        Bank, prepaid court credit and shuttle stock against what members have
        prepaid. Available to all approved members. Dates report the latest recorded
        movement, with the last confirmed physical stock snapshot shown separately.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubPosition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/position:
    get:
      tags: [Ledger]
      summary: The club's true position
      description: >
        Bank, prepaid court credit and shuttle stock against what members have
        prepaid. Legacy admin URL. Approved members use GET /position.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubPosition"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/position/integrity:
    get:
      tags: [Ledger]
      summary: Recompute the club-position identity from the entries
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Integrity report
          content:
            application/json:
              schema:
                type: object
                properties:
                  balanced: { type: boolean }
                  residual_cents: { type: integer, format: int64 }
                  entries_checked: { type: integer }
                  accounts: { type: integer }
        "403":
          $ref: "#/components/responses/Forbidden"

  # --- Settlement ---------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 JWT token
...
  responses:
    TransactionCreated:
      description: Transaction written
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Transaction"
    LedgerConflict:
      description: >
        The request conflicts with what has already happened — the session is
        already settled, the transaction was already reversed, or opening
        balances have already been recorded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LedgerError"
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    LedgerError:
      type: object
      description: >
        A domain failure with a stable machine-readable code, so the frontend can
        react to a specific condition rather than matching on message text.
      properties:
        code:
          type: string
          enum:
            - shuttle_stock_short
            - session_already_settled
            - transaction_already_reversed
            - opening_balances_already_recorded
            - not_settleable
            - invariant_violated
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    PlayerBalance:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        name: { type: string }
        profile_picture: { type: string }
        balance_cents:
          type: integer
          format: int64
          description: Positive means the member is in credit; negative means they owe the club.
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          enum:
            - player_topup
            - withdrawal
            - court_credit_purchase
            - shuttle_purchase
            - session_settlement
            - splitwise_import
            - import_assets
            - opening_balance
            - reversal
        session_id: { type: string, format: uuid }
        reverses_transaction_id: { type: string, format: uuid }
        description: { type: string }
        occurred_at:
          type: string
          format: date-time
          description: When the money moved in the real world; may predate created_at.
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    LedgerActivityView:
      type: object
      required: [id, occurred_at]
      oneOf:
        - required: [entry]
        - required: [game]
      properties:
        id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time, description: Last accounting posting for this activity }
        entry:
          $ref: "#/components/schemas/LedgerEntryView"
        game:
          $ref: "#/components/schemas/LedgerGameView"
    LedgerGameView:
      type: object
      properties:
        session_id: { type: string, format: uuid }
        title: { type: string }
        played_date: { type: string, format: date }
        date_basis: { type: string, enum: [title, recorded] }
        source: { type: string, enum: [splitwise] }
        total_charged_cents: { type: integer, format: int64, description: Sum of gross member shares, including the payer's own share }
        reversed: { type: boolean }
        source_count: { type: integer }
        shares:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              user_id: { type: string, format: uuid }
              member_name: { type: string }
              inactive: { type: boolean }
              charge_cents: { type: integer, format: int64 }
              paid_cents: { type: integer, format: int64, description: Paid on behalf of the group }
              amount_cents: { type: integer, format: int64, description: Net account movement for the game }
              balance_after_cents: { type: integer, format: int64, description: Account balance at the game's last posting }
              guest_names:
                type: array
                items: { type: string }
    LedgerEntryView:
      type: object
      properties:
        id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
        kind: { type: string }
        description: { type: string, description: Original transaction title }
        category:
          type: string
          enum: [topup, session, food, entertainment, shuttles, court_credit, transfer, withdrawal, reversal, opening_balance, other]
        source: { type: string, enum: [splitwise], description: Present only for imported entries }
        member_name: { type: string }
        user_id: { type: string, format: uuid }
        inactive: { type: boolean }
        amount_cents: { type: integer, format: int64 }
        balance_after_cents:
          type: integer
          format: int64
          description: Running balance after this entry, over the member's whole history.
        session_id: { type: string, format: uuid }
        reversed:
          type: boolean
          description: True when a later transaction reversed this one. It stays in the list either way.
    CashRequest:
      type: object
      required: [user_id, amount_cents]
      properties:
        user_id: { type: string, format: uuid }
        amount_cents: { type: integer, format: int64, minimum: 1 }
        occurred_at: { type: string, format: date-time }
        description: { type: string }
    AssetPurchaseRequest:
      type: object
      required: [amount_cents]
      properties:
        amount_cents: { type: integer, format: int64, minimum: 1 }
        units:
          type: integer
          description: Shuttles bought. Required for a shuttle purchase, ignored for court credit.
        occurred_at: { type: string, format: date-time }
        description: { type: string }
    OpeningBalancesRequest:
      type: object
      properties:
        occurred_at: { type: string, format: date-time }
        players:
          type: array
          items:
            type: object
            properties:
              user_id: { type: string, format: uuid }
              balance_cents: { type: integer, format: int64 }
        bank_cents: { type: integer, format: int64 }
        court_credit_cents: { type: integer, format: int64 }
        shuttle_stock:
          type: object
          properties:
            units: { type: integer }
            amount_cents: { type: integer, format: int64 }
    ClubPosition:
      type: object
      properties:
        assets_pending:
          type: boolean
          description: Imported member history is verified but the actual asset snapshot has not been supplied. Do not present placeholder asset values as confirmed holdings.
        assets:
          type: object
          properties:
            bank_cents: { type: integer, format: int64 }
            court_credit_cents: { type: integer, format: int64 }
            shuttle_stock_cents: { type: integer, format: int64 }
            shuttle_stock_units: { type: integer }
            total_cents: { type: integer, format: int64 }
            bank_as_of: { type: string, format: date, nullable: true }
            court_credit_as_of: { type: string, format: date, nullable: true }
            shuttle_stock_as_of: { type: string, format: date, nullable: true }
            shuttle_audited_on:
              type: string
              format: date
              nullable: true
              description: Last unreversed opening or verified import stock snapshot. Purchases and consumption do not advance this date.
        liabilities:
          type: object
          properties:
            player_balances_cents: { type: integer, format: int64 }
        surplus_cents: { type: integer, format: int64 }
        balanced:
          type: boolean
          description: >
            Restates the club-position identity. Always true in normal operation,
            since a posting that would falsify it is rolled back — so false means
            something wrote entries without going through the ledger service.
        warnings:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              next_session_id: { type: string, format: uuid }

    # --- Settlement ---------------------------------------------------------
...